changing short to int

+++ Ex02/SecretWordGenerator.cs
@@ -7,14 +7,14 @@ namespace Ex02
     {
         private readonly Random r_RandomGenerator = new Random();
         private readonly char[] r_CharsBank = {'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H'};
-        private const short k_SecretWordLength = 4;     //remember CONST MEANS #DEFINE IN CSHARP
+        private const int k_SecretWordLength = 4;     //remember CONST MEANS #DEFINE IN CSHARP
 
         public string GenerateSecretWord()
         {
             string secretWord = string.Empty;
             List<char> availableChars = new List<char>(r_CharsBank);
 
-            for (short i = 0; i < k_SecretWordLength; i++)
+            for (int i = 0; i < k_SecretWordLength; i++)
             {
                 int randomIndex = r_RandomGenerator.Next(availableChars.Count);
                 secretWord += availableChars[randomIndex];
